+++ src/main/java/com/transport/management/modules/conductor/ConductorController.java
@@ -12,8 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.transport.management.services.ConductorService;
-
 @RestController
 @RequestMapping("/conductores")
 public class ConductorController {
